+++ cmd/template/cluster/provider/templates/aws/template.go
@@ -7,6 +7,9 @@ import (
 //go:embed cluster.yaml.tmpl
 var cluster string
 
+//go:embed cluster_eks.yaml.tmpl
+var clusterEKS string
+
 //go:embed aws_cluster.yaml.tmpl
 var awsCluster string
 
@@ -15,6 +18,9 @@ var kubeadmControlPlane string
 
 //go:embed aws_machine_template.yaml.tmpl
 var awsMachineTemplate string
+
+//go:embed aws_managed_control_plane.yaml.tmpl
+var awsManagedControlPlane string
 
 //go:embed aws_cluster_role_identity.yaml.tmpl
 var awsClusterRoleIdentity string
@@ -33,8 +39,8 @@ type Template struct {
 	Data string
 }
 
-// GetTemplate merges all .tmpl files.
-func GetTemplates() []Template {
+// GetAWSTemplate merges .tmpl files for an AWS cluster.
+func GetAWSTemplates() []Template {
 	// Order is important here.
 	// The order in this slice determines in which order files will be applied.
 	return []Template{
@@ -46,5 +52,16 @@ func GetTemplates() []Template {
 		{Name: "bastion_secret.yaml.tmpl", Data: bastionSecret},
 		{Name: "bastion_machine_deployment.yaml.tmpl", Data: bastionMachineDeployment},
 		{Name: "bastion_aws_machine_template.yaml.tmpl", Data: bastionAWSMachineTemplate},
+	}
+}
+
+// GetEKSTemplate merges .tmpl files for an EKS cluster.
+func GetEKSTemplates() []Template {
+	// Order is important here.
+	// The order in this slice determines in which order files will be applied.
+	return []Template{
+		{Name: "cluster_eks.yaml.tmpl", Data: clusterEKS},
+		{Name: "aws_managed_control_plane.yaml.tmpl", Data: awsManagedControlPlane},
+		{Name: "aws_cluster_role_identity.yaml.tmpl", Data: awsClusterRoleIdentity},
 	}
 }